+++ ktlint-gradle-plugin/src/main/kotlin/com/rickbusarow/ktlint/KtLintPlugin.kt
@@ -81,16 +81,23 @@ abstract class KtLintPlugin : Plugin<GradleProject> {
       sourceFileShadowDirectory = target.sourceFileShadowDirectory("gradleScripts"),
       sourceDirectorySet = target.provider {
 
+        val includedBuildDirs = target.gradle.includedBuilds.map { it.projectDir }
         val subProjectDirs = target.subprojects.mapToSet { it.projectDir }
         val srcDirs = sourceSets.values.flatMapToSet { it.get().kotlin.srcDirs }
         val reg = Regex(""".*\.gradle\.kts$""")
         target.files(
           target.projectDir
             .walkTopDown()
-            .onEnter { it != target.buildDir }
-            .onEnter { it != target.file("src") }
-            .onEnter { it !in subProjectDirs }
-            .onEnter { it !in srcDirs }
+            .onEnter {
+              when (it) {
+                target.buildDir -> false
+                target.file("src") -> false
+                in includedBuildDirs -> false
+                in subProjectDirs -> false
+                in srcDirs -> false
+                else -> true
+              }
+            }
             .filter { it.name.matches(reg) }.toList()
         )
       },